+++ spec/resolver.spec.js
@@ -3,14 +3,58 @@ var bindingsResolver = postal.configuration.resolver;
 
 describe( "amqpBindingsResolver", function() {
 	describe( "When calling compare", function() {
-		before( function() {
+		beforeEach( function() {
 			bindingsResolver.reset();
+		} );
+		describe( "and caching comparisons", function() {
+			it( "should cache a passing comparison", function() {
+				bindingsResolver.compare( "test", "test" );
+				bindingsResolver.cache.should.have.ownProperty( "test|test" );
+				bindingsResolver.cache[ "test|test" ].should.be.ok;
+			} );
+			it( "should cache a failing comparison", function() {
+				bindingsResolver.compare( "nope", "test" );
+				bindingsResolver.cache.should.have.ownProperty( "test|nope" );
+				bindingsResolver.cache[ "test|nope" ].should.not.be.ok;
+			} );
+			it( "should purge the cache when calling purge with no arguments", function() {
+				bindingsResolver.compare( "test", "test" );
+				bindingsResolver.compare( "nope", "test" );
+				bindingsResolver.cache.should.have.ownProperty( "test|test" );
+				bindingsResolver.cache.should.have.ownProperty( "test|nope" );
+				bindingsResolver.purge();
+				( bindingsResolver.cache ).should.be.empty;
+			} );
+			it( "should purge the topics specified when calling purge", function() {
+				bindingsResolver.compare( "test", "test" );
+				bindingsResolver.compare( "nope", "test" );
+				bindingsResolver.compare( "test", "nope" );
+				bindingsResolver.cache.should.have.ownProperty( "test|test" );
+				bindingsResolver.cache.should.have.ownProperty( "test|nope" );
+				bindingsResolver.cache.should.have.ownProperty( "nope|test" );
+				bindingsResolver.purge( { topic: "test" } );
+				( bindingsResolver.cache ).should.have.ownProperty( "nope|test" );
+				bindingsResolver.cache.should.not.have.ownProperty( "test|test" );
+				bindingsResolver.cache.should.not.have.ownProperty( "test|nope" );
+			} );
+			it( "should purge the bindings specified when calling purge", function() {
+				bindingsResolver.compare( "test", "test" );
+				bindingsResolver.compare( "nope", "test" );
+				bindingsResolver.compare( "test", "nope" );
+				bindingsResolver.cache.should.have.ownProperty( "test|test" );
+				bindingsResolver.cache.should.have.ownProperty( "test|nope" );
+				bindingsResolver.cache.should.have.ownProperty( "nope|test" );
+				bindingsResolver.purge( { binding: "test" } );
+				( bindingsResolver.cache ).should.have.ownProperty( "test|nope" );
+				bindingsResolver.cache.should.not.have.ownProperty( "test|test" );
+				bindingsResolver.cache.should.not.have.ownProperty( "nope|test" );
+			} );
 		} );
 		describe( "With '*' wildcards", function() {
 			// Passing matches
 			describe( "With topic Top.Middle.Bottom and binding *.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "*.Middle.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-*.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|*.Middle.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -20,7 +64,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding Top.*.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.*.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.*.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -30,7 +74,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding Top.Middle.*", function() {
 				var result = bindingsResolver.compare( "Top.Middle.*", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.Middle.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.Middle.*" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -40,7 +84,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding Top.*.*", function() {
 				var result = bindingsResolver.compare( "Top.*.*", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.*.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.*.*" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -50,7 +94,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding Top.*.*", function() {
 				var result = bindingsResolver.compare( "*.*.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-*.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|*.*.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -60,7 +104,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding *.Middle.*", function() {
 				var result = bindingsResolver.compare( "*.Middle.*", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-*.Middle.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|*.Middle.*" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -70,7 +114,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding *.*.*", function() {
 				var result = bindingsResolver.compare( "*.*.*", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-*.*.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|*.*.*" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -82,27 +126,27 @@ describe( "amqpBindingsResolver", function() {
 			// Failing Matches
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding *.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "*.Middle.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-*.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|*.Middle.Bottom" ];
 				it( "Result should be false", function() {
-					result.should.not.be.ok
+					result.should.not.be.ok;
 				} );
 				it( "Should *not* create a resolver cache entry", function() {
-					cached.should.not.be.ok
+					cached.should.not.be.ok;
 				} );
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding Top.*.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.*.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-Top.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|Top.*.Bottom" ];
 				it( "Result should be false", function() {
-					result.should.not.be.ok
+					result.should.not.be.ok;
 				} );
 				it( "Should *not* create a resolver cache entry", function() {
-					cached.should.not.be.ok
+					cached.should.not.be.ok;
 				} );
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding Top.Middle.*", function() {
 				var result = bindingsResolver.compare( "Top.Middle.*", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-Top.Middle.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|Top.Middle.*" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -112,7 +156,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding Top.*.*", function() {
 				var result = bindingsResolver.compare( "Top.*.*", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-Top.*.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|Top.*.*" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -122,7 +166,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding Top.*.*", function() {
 				var result = bindingsResolver.compare( "*.*.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-*.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|*.*.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -132,7 +176,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding *.Middle.*", function() {
 				var result = bindingsResolver.compare( "*.Middle.*", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-*.Middle.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|*.Middle.*" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -142,7 +186,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding *.*.*", function() {
 				var result = bindingsResolver.compare( "*.*.*", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-*.*.*" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|*.*.*" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -150,14 +194,13 @@ describe( "amqpBindingsResolver", function() {
 					cached.should.not.be.ok;
 				} );
 			} );
-
 		} );
 		describe( "With '#' wildcards", function() {
 			// Passing matches
 			// # at beginning of binding
 			describe( "With topic Top.Middle.Bottom and binding #.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.Middle.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|#.Middle.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -167,7 +210,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.SubTop.Middle.Bottom and binding #.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.Middle.Bottom", "Top.SubTop.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.SubTop.Middle.Bottom-#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.SubTop.Middle.Bottom|#.Middle.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -177,7 +220,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Middle.Bottom and binding #.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.Middle.Bottom", "Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Middle.Bottom-#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Middle.Bottom|#.Middle.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -188,7 +231,7 @@ describe( "amqpBindingsResolver", function() {
 			// # in middle of binding
 			describe( "With topic Top.Middle.Bottom and binding Top.#.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.#.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.#.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -198,7 +241,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding Top.#.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.#.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-Top.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|Top.#.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -208,7 +251,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.SubTop.Middle.SubMiddle.Bottom and binding Top.#.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.#.Bottom", "Top.SubTop.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.SubTop.Middle.SubMiddle.Bottom-Top.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.SubTop.Middle.SubMiddle.Bottom|Top.#.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -218,7 +261,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Bottom and binding Top.#.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.#.Bottom", "Top.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Bottom-Top.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Bottom|Top.#.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -229,7 +272,7 @@ describe( "amqpBindingsResolver", function() {
 			// # at end of binding
 			describe( "With topic Top.Middle.Bottom and binding Top.Middle.#", function() {
 				var result = bindingsResolver.compare( "Top.Middle.#", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.Middle.#" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.Middle.#" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -239,7 +282,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.SubTop.Middle.Bottom and binding Top.SubTop.#", function() {
 				var result = bindingsResolver.compare( "Top.SubTop.#", "Top.SubTop.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.SubTop.Middle.Bottom-Top.SubTop.#" ];
+					cached = bindingsResolver.cache[ "Top.SubTop.Middle.Bottom|Top.SubTop.#" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -249,7 +292,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Middle.Bottom and binding Middle.#", function() {
 				var result = bindingsResolver.compare( "Middle.#", "Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Middle.Bottom-Middle.#" ];
+					cached = bindingsResolver.cache[ "Middle.Bottom|Middle.#" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -261,7 +304,7 @@ describe( "amqpBindingsResolver", function() {
 			// # at beginning of binding
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding #.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.Middle.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|#.Middle.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -271,7 +314,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.SubTop.Middle.SubMiddle.Bottom and binding #.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.Middle.Bottom", "Top.SubTop.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.SubTop.Middle.SubMiddle.Bottom-#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.SubTop.Middle.SubMiddle.Bottom|#.Middle.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -281,7 +324,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Middle.Bottom and binding #.Middle.SubMiddle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.Middle.Bottom", "Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Middle.SubMiddle.Bottom-#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Middle.SubMiddle.Bottom|#.Middle.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -292,7 +335,7 @@ describe( "amqpBindingsResolver", function() {
 			// # in middle of binding
 			describe( "With topic Top.Middle.Bottom and binding Top.SubTop.#.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.SubTop.#.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.SubTop.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.SubTop.#.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -302,7 +345,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom.SubBottom and binding Top.#.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.#.Bottom", "Top.Middle.SubMiddle.Bottom.SubBottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom.SubBottom-Top.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom.SubBottom|Top.#.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -312,7 +355,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.SubTop.Middle.SubMiddle.Bottom and binding Top.#.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.#.Middle.Bottom", "Top.SubTop.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.SubTop.Middle.SubMiddle.Bottom-Top.#.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.SubTop.Middle.SubMiddle.Bottom|Top.#.Middle.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -322,7 +365,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.SubTop.Bottom and binding SubTop.#.Bottom", function() {
 				var result = bindingsResolver.compare( "SubTop.#.Bottom", "Top.SubTop.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.SubTop.Bottom-SubTop.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.SubTop.Bottom|SubTop.#.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -333,7 +376,7 @@ describe( "amqpBindingsResolver", function() {
 			// # at end of binding
 			describe( "With topic Top.Bottom and binding Top.Middle.#", function() {
 				var result = bindingsResolver.compare( "Top.Middle.#", "Top.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Bottom-Top.Middle.#" ];
+					cached = bindingsResolver.cache[ "Top.Bottom|Top.Middle.#" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -343,7 +386,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.Bottom and binding Top.SubTop.#", function() {
 				var result = bindingsResolver.compare( "Top.SubTop.#", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.SubTop.#" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.SubTop.#" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -353,7 +396,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Bottom and binding Middle.#", function() {
 				var result = bindingsResolver.compare( "Middle.#", "Bottom" ),
-					cached = bindingsResolver.cache[ "Bottom-Middle.#" ];
+					cached = bindingsResolver.cache[ "Bottom|Middle.#" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -366,7 +409,7 @@ describe( "amqpBindingsResolver", function() {
 			// Passing matches
 			describe( "With topic Top.Middle.Bottom and binding #.*.Bottom", function() {
 				var result = bindingsResolver.compare( "#.*.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-#.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|#.*.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -376,7 +419,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding #.*.Bottom", function() {
 				var result = bindingsResolver.compare( "#.*.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-#.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|#.*.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -386,7 +429,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Bottom and binding #.*.Bottom", function() {
 				var result = bindingsResolver.compare( "#.*.Bottom", "Top.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Bottom-#.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Bottom|#.*.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -396,7 +439,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Bottom and binding *.#.Bottom", function() {
 				var result = bindingsResolver.compare( "*.#.Bottom", "Top.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Bottom-*.#.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Bottom|*.#.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -407,7 +450,7 @@ describe( "amqpBindingsResolver", function() {
 			// Failing matches
 			describe( "With topic Bottom and binding #.*.Bottom", function() {
 				var result = bindingsResolver.compare( "#.*.Bottom", "Bottom" ),
-					cached = bindingsResolver.cache[ "Bottom-#.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Bottom|#.*.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -417,7 +460,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Middle.SubMiddle.Bottom and binding Top.Middle.SubMiddle.#.*.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.Middle.SubMiddle.#.*.Bottom", "Top.Middle.SubMiddle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom-Top.Middle.SubMiddle.#.*.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.SubMiddle.Bottom|Top.Middle.SubMiddle.#.*.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -427,7 +470,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic Top.Bottom and binding #.*.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "#.*.Middle.Bottom", "Top.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Bottom-#.*.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Bottom|#.*.Middle.Bottom" ];
 				it( "Result should be false", function() {
 					result.should.not.be.ok;
 				} );
@@ -439,7 +482,7 @@ describe( "amqpBindingsResolver", function() {
 		describe( "With plain string matching", function() {
 			describe( "With topic Top.Middle.Bottom and binding Top.Middle.Bottom", function() {
 				var result = bindingsResolver.compare( "Top.Middle.Bottom", "Top.Middle.Bottom" ),
-					cached = bindingsResolver.cache[ "Top.Middle.Bottom-Top.Middle.Bottom" ];
+					cached = bindingsResolver.cache[ "Top.Middle.Bottom|Top.Middle.Bottom" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -449,7 +492,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic 'Topic' and binding 'Topic'", function() {
 				var result = bindingsResolver.compare( "Topic", "Topic" ),
-					cached = bindingsResolver.cache[ "Topic-Topic" ];
+					cached = bindingsResolver.cache[ "Topic|Topic" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );
@@ -459,7 +502,7 @@ describe( "amqpBindingsResolver", function() {
 			} );
 			describe( "With topic '/sample/topic' and binding '/sample/topic'", function() {
 				var result = bindingsResolver.compare( "/sample/topic", "/sample/topic" ),
-					cached = bindingsResolver.cache[ "/sample/topic-/sample/topic" ];
+					cached = bindingsResolver.cache[ "/sample/topic|/sample/topic" ];
 				it( "Result should be true", function() {
 					result.should.be.ok;
 				} );