+++ spec/utils.spec.js
@@ -1,4 +1,4 @@
-/* global postal, _ */
+/* global global, postal, _ */
 var NO_OP = function() {};
 describe( "postal.utils", function() {
 	beforeEach( function() {
@@ -60,6 +60,7 @@ describe( "postal.utils", function() {
 				res += 1;
 			};
 			beforeEach( function() {
+				postal.reset();
 				subs.push( postal.subscribe( {
 					channel: "A",
 					topic: "some.topic",
@@ -253,6 +254,105 @@ describe( "postal.utils", function() {
 			} );
 			it( "should have not invoked subscriber callbacks when publishing", function() {
 				res.should.equal( 1 );
+			} );
+		} );
+		describe( "with a custom string property", function() {
+			var subs = [];
+			var res = 0;
+			var cb = function() {
+				res += 1;
+			};
+			beforeEach( function() {
+				subs.push( _.extend( postal.subscribe( {
+					channel: "A",
+					topic: "some.topic",
+					callback: cb
+				} ), { thingy: "some/string/[value]" } ) );
+				subs.push( _.extend( postal.subscribe( {
+					channel: "B",
+					topic: "another.topic",
+					callback: cb
+				} ), { thingy: "some/string/[value]" } ) );
+				subs.push( _.extend( postal.subscribe( {
+					channel: "B",
+					topic: "even.more.topics",
+					callback: cb
+				} ), { thingy: "some/string/[value]" } ) );
+				postal.unsubscribeFor( { channel: "B", thingy: "some/string/[value]" } );
+				postal.publish( {
+					channel: "B",
+					topic: "another.topic",
+					data: {}
+				} );
+				postal.publish( {
+					channel: "B",
+					topic: "even.more.topics",
+					data: {}
+				} );
+			} );
+			afterEach( function() {
+				res = 0;
+				subs = [];
+				postal.reset();
+			} );
+			it( "should have removed correct subscribers", function() {
+				( typeof postal.subscriptions.B === "undefined" ).should.be.ok;
+				( typeof postal.subscriptions.B === "undefined" ).should.be.ok;
+				( postal.subscriptions.A[ "some.topic" ] ).should.be.ok;
+				postal.getSubscribersFor( { channel: "B", thingy: "some/string/[value]" } ).length.should.equal( 0 );
+			} );
+			it( "should have not invoked subscriber callbacks when publishing", function() {
+				res.should.equal( 0 );
+			} );
+		} );
+		describe( "with a custom object property", function() {
+			var subs = [];
+			var res = 0;
+			var cb = function() {
+				res += 1;
+			};
+			var objectyObj = { greeting: "oh, hai" };
+			beforeEach( function() {
+				subs.push( _.extend( postal.subscribe( {
+					channel: "A",
+					topic: "some.topic",
+					callback: cb
+				} ), { thingy: objectyObj } ) );
+				subs.push( _.extend( postal.subscribe( {
+					channel: "B",
+					topic: "another.topic",
+					callback: cb
+				} ), { thingy: objectyObj } ) );
+				subs.push( _.extend( postal.subscribe( {
+					channel: "B",
+					topic: "even.more.topics",
+					callback: cb
+				} ), { thingy: objectyObj } ) );
+				postal.unsubscribeFor( { channel: "B", thingy: objectyObj } );
+				postal.publish( {
+					channel: "B",
+					topic: "another.topic",
+					data: {}
+				} );
+				postal.publish( {
+					channel: "B",
+					topic: "even.more.topics",
+					data: {}
+				} );
+			} );
+			afterEach( function() {
+				res = 0;
+				subs = [];
+				postal.reset();
+			} );
+			it( "should have removed correct subscribers", function() {
+				( typeof postal.subscriptions.B === "undefined" ).should.be.ok;
+				( typeof postal.subscriptions.B === "undefined" ).should.be.ok;
+				( postal.subscriptions.A[ "some.topic" ] ).should.be.ok;
+				postal.getSubscribersFor( { channel: "B", thingy: objectyObj } ).length.should.equal( 0 );
+			} );
+			it( "should have not invoked subscriber callbacks when publishing", function() {
+				res.should.equal( 0 );
 			} );
 		} );
 	} );